Refactor data capture -- single patch
