+++ data/collector.go
@@ -139,6 +139,10 @@ func (c *collector) Collect() {
 		defer c.logRoutine.Done()
 		c.logCaptureErrs()
 	})
+
+	// We must wait on `started` before returning. The sleep/ticker based captures rely on the clock
+	// advancing to do their first "tick". They must make an initial clock reading before unittests
+	// add an "interval". Lest the ticker never fires and a reading is never made.
 	<-started
 }
