+++ services/datamanager/builtin/builtin.go
@@ -3,7 +3,6 @@ package builtin
 
 import (
 	"context"
-	"fmt"
 	"net"
 	"os"
 	"path/filepath"
@@ -22,7 +21,6 @@ import (
 	"go.viam.com/rdk/data"
 	"go.viam.com/rdk/internal/cloud"
 	"go.viam.com/rdk/logging"
-	"go.viam.com/rdk/protoutils"
 	"go.viam.com/rdk/resource"
 	"go.viam.com/rdk/services/datamanager"
 	"go.viam.com/rdk/services/datamanager/datacapture"
@@ -44,21 +42,8 @@ func init() {
 		})
 }
 
-// TODO: re-determine if queue size is optimal given we now support 10khz+ capture rates
-// The Collector's queue should be big enough to ensure that .capture() is never blocked by the queue being
-// written to disk. A default value of 250 was chosen because even with the fastest reasonable capture interval (1ms),
-// this would leave 250ms for a (buffered) disk write before blocking, which seems sufficient for the size of
-// writes this would be performing.
-const defaultCaptureQueueSize = 250
-
-// Default bufio.Writer buffer size in bytes.
-const defaultCaptureBufferSize = 4096
-
 // Default time to wait in milliseconds to check if a file has been modified.
 const defaultFileLastModifiedMillis = 10000.0
-
-// Default maximum size in bytes of a data capture file.
-var defaultMaxCaptureSize = int64(256 * 1024)
 
 // Default time between disk size checks.
 var filesystemPollInterval = 30 * time.Second
@@ -73,8 +58,6 @@ var (
 	clock          = clk.New()
 	deletionTicker = clk.New()
 )
-
-var errCaptureDirectoryConfigurationDisabled = errors.New("changing the capture directory is prohibited in this environment")
 
 // Config describes how to configure the service.
 type Config struct {
@@ -127,14 +110,11 @@ func readyToSync(ctx context.Context, s selectiveSyncer, logger logging.Logger) 
 type builtIn struct {
 	resource.Named
 	logger                 logging.Logger
-	captureDir             string
-	captureDisabled        bool
-	collectorsMu           sync.Mutex
-	collectors             map[resourceMethodMetadata]*collectorAndConfig
 	lock                   sync.Mutex
 	backgroundWorkers      sync.WaitGroup
 	fileLastModifiedMillis int
 
+	// Dan: This now includes the capture dir. Change to syncPaths
 	additionalSyncPaths []string
 	tags                []string
 	syncDisabled        bool
@@ -147,21 +127,15 @@ type builtIn struct {
 	cloudConnSvc        cloud.ConnectionService
 	cloudConn           rpc.ClientConn
 	syncTicker          *clk.Ticker
-	maxCaptureFileSize  int64
 
 	syncSensor           selectiveSyncer
 	selectiveSyncEnabled bool
 
-	componentMethodFrequencyHz map[resourceMethodMetadata]float32
-
 	fileDeletionRoutineCancelFn   context.CancelFunc
 	fileDeletionBackgroundWorkers *sync.WaitGroup
 
-	captureDirPollingCancelFn          context.CancelFunc
-	captureDirPollingBackgroundWorkers *sync.WaitGroup
+	captureManager *data.CaptureManager
 }
-
-var viamCaptureDotDir = filepath.Join(os.Getenv("HOME"), ".viam", "capture")
 
 // NewBuiltIn returns a new data manager service for the given robot.
 func NewBuiltIn(
@@ -171,17 +145,15 @@ func NewBuiltIn(
 	logger logging.Logger,
 ) (datamanager.Service, error) {
 	svc := &builtIn{
-		Named:                      conf.ResourceName().AsNamed(),
-		logger:                     logger,
-		captureDir:                 viamCaptureDotDir,
-		collectors:                 make(map[resourceMethodMetadata]*collectorAndConfig),
-		syncIntervalMins:           0,
-		additionalSyncPaths:        []string{},
-		tags:                       []string{},
-		fileLastModifiedMillis:     defaultFileLastModifiedMillis,
-		syncerConstructor:          datasync.NewManager,
-		selectiveSyncEnabled:       false,
-		componentMethodFrequencyHz: make(map[resourceMethodMetadata]float32),
+		Named:                  conf.ResourceName().AsNamed(),
+		logger:                 logger,
+		syncIntervalMins:       0,
+		additionalSyncPaths:    []string{},
+		tags:                   []string{},
+		fileLastModifiedMillis: defaultFileLastModifiedMillis,
+		syncerConstructor:      datasync.NewManager,
+		selectiveSyncEnabled:   false,
+		captureManager:         data.NewCaptureManager(logger.Sublogger("capture"), clock),
 	}
 
 	if err := svc.Reconfigure(ctx, deps, conf); err != nil {
@@ -194,7 +166,6 @@ func NewBuiltIn(
 // Close releases all resources managed by data_manager.
 func (svc *builtIn) Close(_ context.Context) error {
 	svc.lock.Lock()
-	svc.closeCollectors()
 	svc.closeSyncer()
 	if svc.syncRoutineCancelFn != nil {
 		svc.syncRoutineCancelFn()
@@ -202,189 +173,17 @@ func (svc *builtIn) Close(_ context.Context) error {
 	if svc.fileDeletionRoutineCancelFn != nil {
 		svc.fileDeletionRoutineCancelFn()
 	}
-	if svc.captureDirPollingCancelFn != nil {
-		svc.captureDirPollingCancelFn()
-	}
 
 	fileDeletionBackgroundWorkers := svc.fileDeletionBackgroundWorkers
-	capturePollingWorker := svc.captureDirPollingBackgroundWorkers
 	svc.lock.Unlock()
+	svc.captureManager.Close()
 	svc.backgroundWorkers.Wait()
 
 	if fileDeletionBackgroundWorkers != nil {
 		fileDeletionBackgroundWorkers.Wait()
 	}
-	if capturePollingWorker != nil {
-		capturePollingWorker.Wait()
-	}
 
 	return nil
-}
-
-func (svc *builtIn) closeCollectors() {
-	var collectorsToClose []data.Collector
-	svc.collectorsMu.Lock()
-	for _, collectorAndConfig := range svc.collectors {
-		collectorsToClose = append(collectorsToClose, collectorAndConfig.Collector)
-	}
-	svc.collectors = make(map[resourceMethodMetadata]*collectorAndConfig)
-	svc.collectorsMu.Unlock()
-
-	var wg sync.WaitGroup
-	for _, collector := range collectorsToClose {
-		wg.Add(1)
-		go func(toClose data.Collector) {
-			defer wg.Done()
-			toClose.Close()
-		}(collector)
-	}
-	wg.Wait()
-}
-
-func (svc *builtIn) flushCollectors() {
-	var collectorsToFlush []data.Collector
-	svc.collectorsMu.Lock()
-	for _, collectorAndConfig := range svc.collectors {
-		collectorsToFlush = append(collectorsToFlush, collectorAndConfig.Collector)
-	}
-	svc.collectorsMu.Unlock()
-
-	var wg sync.WaitGroup
-	for _, collector := range collectorsToFlush {
-		wg.Add(1)
-		go func(toFlush data.Collector) {
-			defer wg.Done()
-			toFlush.Flush()
-		}(collector)
-	}
-	wg.Wait()
-}
-
-// Parameters stored for each collector.
-type collectorAndConfig struct {
-	Resource  resource.Resource
-	Collector data.Collector
-	Config    datamanager.DataCaptureConfig
-}
-
-// Identifier for a particular collector: component name, component model, component type,
-// method parameters, and method name.
-type resourceMethodMetadata struct {
-	ResourceName   string
-	MethodParams   string
-	MethodMetadata data.MethodMetadata
-}
-
-func (r resourceMethodMetadata) String() string {
-	return fmt.Sprintf(
-		"[API: %s, Resource Name: %s, Method Name: %s, Method Params: %s]",
-		r.MethodMetadata.API, r.ResourceName, r.MethodMetadata.MethodName, r.MethodParams)
-}
-
-// Get time.Duration from hz.
-func getDurationFromHz(captureFrequencyHz float32) time.Duration {
-	if captureFrequencyHz == 0 {
-		return time.Duration(0)
-	}
-	return time.Duration(float32(time.Second) / captureFrequencyHz)
-}
-
-var metadataToAdditionalParamFields = map[string]string{
-	generateMetadataKey("rdk:component:board", "Analogs"): "reader_name",
-	generateMetadataKey("rdk:component:board", "Gpios"):   "pin_name",
-}
-
-// Initialize a collector for the component/method or update it if it has previously been created.
-// Return the component/method metadata which is used as a key in the collectors map.
-func (svc *builtIn) initializeOrUpdateCollector(
-	res resource.Resource,
-	md resourceMethodMetadata,
-	config datamanager.DataCaptureConfig,
-	maxFileSizeChanged bool,
-) (*collectorAndConfig, error) {
-	// Build metadata.
-	captureMetadata, err := datacapture.BuildCaptureMetadata(
-		config.Name.API,
-		config.Name.ShortName(),
-		config.Method,
-		config.AdditionalParams,
-		config.Tags,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	// TODO(DATA-451): validate method params
-
-	svc.collectorsMu.Lock()
-	defer svc.collectorsMu.Unlock()
-	if storedCollectorAndConfig, ok := svc.collectors[md]; ok {
-		if storedCollectorAndConfig.Config.Equals(&config) &&
-			res == storedCollectorAndConfig.Resource &&
-			!maxFileSizeChanged {
-			// If the attributes have not changed, do nothing and leave the existing collector.
-			return svc.collectors[md], nil
-		}
-		// If the attributes have changed, close the existing collector.
-		storedCollectorAndConfig.Collector.Close()
-	}
-
-	// Get collector constructor for the component API and method.
-	collectorConstructor := data.CollectorLookup(md.MethodMetadata)
-	if collectorConstructor == nil {
-		return nil, errors.Errorf("failed to find collector constructor for %s", md.MethodMetadata)
-	}
-
-	// Parameters to initialize collector.
-	interval := getDurationFromHz(config.CaptureFrequencyHz)
-	// Set queue size to defaultCaptureQueueSize if it was not set in the config.
-	captureQueueSize := config.CaptureQueueSize
-	if captureQueueSize == 0 {
-		captureQueueSize = defaultCaptureQueueSize
-	}
-
-	captureBufferSize := config.CaptureBufferSize
-	if captureBufferSize == 0 {
-		captureBufferSize = defaultCaptureBufferSize
-	}
-	additionalParamKey, ok := metadataToAdditionalParamFields[generateMetadataKey(
-		md.MethodMetadata.API.String(),
-		md.MethodMetadata.MethodName)]
-	if ok {
-		if _, ok := config.AdditionalParams[additionalParamKey]; !ok {
-			return nil, errors.Errorf("failed to validate additional_params for %s, must supply %s",
-				md.MethodMetadata.API.String(), additionalParamKey)
-		}
-	}
-	methodParams, err := protoutils.ConvertStringMapToAnyPBMap(config.AdditionalParams)
-	if err != nil {
-		return nil, err
-	}
-
-	// Create a collector for this resource and method.
-	targetDir := datacapture.FilePathWithReplacedReservedChars(
-		filepath.Join(svc.captureDir, captureMetadata.GetComponentType(),
-			captureMetadata.GetComponentName(), captureMetadata.GetMethodName()))
-	if err := os.MkdirAll(targetDir, 0o700); err != nil {
-		return nil, err
-	}
-	params := data.CollectorParams{
-		ComponentName: config.Name.ShortName(),
-		Interval:      interval,
-		MethodParams:  methodParams,
-		Target:        datacapture.NewBuffer(targetDir, captureMetadata, svc.maxCaptureFileSize),
-		QueueSize:     captureQueueSize,
-		BufferSize:    captureBufferSize,
-		Logger:        svc.logger,
-		Clock:         clock,
-	}
-	collector, err := (*collectorConstructor)(res, params)
-	if err != nil {
-		return nil, err
-	}
-	collector.Collect()
-
-	return &collectorAndConfig{res, collector, config}, nil
 }
 
 func (svc *builtIn) closeSyncer() {
@@ -414,8 +213,7 @@ func (svc *builtIn) initSyncer(ctx context.Context) error {
 	}
 
 	client := v1.NewDataSyncServiceClient(conn)
-	svc.filesToSync = make(chan string, svc.maxSyncThreads)
-	syncer, err := svc.syncerConstructor(identity, client, svc.logger, svc.captureDir, svc.maxSyncThreads, svc.filesToSync)
+	syncer, err := svc.syncerConstructor(identity, client, svc.logger, svc.captureManager.CaptureDir(), svc.maxSyncThreads)
 	if err != nil {
 		return errors.Wrap(err, "failed to initialize new syncer")
 	}
@@ -463,6 +261,17 @@ func (svc *builtIn) Reconfigure(
 		return err
 	}
 
+	dataConfig := data.Config{
+		CaptureDisabled:             svcConfig.CaptureDisabled,
+		CaptureDir:                  svcConfig.CaptureDir,
+		Tags:                        svcConfig.Tags,
+		MaximumCaptureFileSizeBytes: svcConfig.MaximumCaptureFileSizeBytes,
+	}
+	if err = svc.captureManager.Reconfigure(ctx, deps, conf, dataConfig); err != nil {
+		svc.logger.Warnw("DataCapture reconfigure error", "err", err)
+		return err
+	}
+
 	// Syncer should be reinitialized if the max sync threads are updated in the config
 	newMaxSyncThreadValue := datasync.MaxParallelSyncRoutines
 	if svcConfig.MaximumNumSyncThreads != 0 {
@@ -470,39 +279,6 @@ func (svc *builtIn) Reconfigure(
 	}
 	reinitSyncer := cloudConnSvc != svc.cloudConnSvc || newMaxSyncThreadValue != svc.maxSyncThreads
 	svc.cloudConnSvc = cloudConnSvc
-
-	captureConfigs, err := svc.updateDataCaptureConfigs(deps, conf, svcConfig.CaptureDir)
-	if err != nil {
-		return err
-	}
-
-	if !utils.IsTrustedEnvironment(ctx) && svcConfig.CaptureDir != "" && svcConfig.CaptureDir != viamCaptureDotDir {
-		return errCaptureDirectoryConfigurationDisabled
-	}
-
-	if svcConfig.CaptureDir != "" {
-		svc.captureDir = svcConfig.CaptureDir
-	} else {
-		svc.captureDir = viamCaptureDotDir
-	}
-
-	if svc.captureDirPollingCancelFn != nil {
-		svc.captureDirPollingCancelFn()
-	}
-	if svc.captureDirPollingBackgroundWorkers != nil {
-		svc.captureDirPollingBackgroundWorkers.Wait()
-	}
-	captureDirPollCtx, captureDirCancelFunc := context.WithCancel(context.Background())
-	svc.captureDirPollingCancelFn = captureDirCancelFunc
-	svc.captureDirPollingBackgroundWorkers = &sync.WaitGroup{}
-	svc.captureDirPollingBackgroundWorkers.Add(1)
-	go logCaptureDirSize(captureDirPollCtx, svc.captureDir, svc.captureDirPollingBackgroundWorkers, svc.logger)
-
-	svc.captureDisabled = svcConfig.CaptureDisabled
-	// Service is disabled, so close all collectors and clear the map so we can instantiate new ones if we enable this service.
-	if svc.captureDisabled {
-		svc.closeCollectors()
-	}
 
 	if svc.fileDeletionRoutineCancelFn != nil {
 		svc.fileDeletionRoutineCancelFn()
@@ -515,78 +291,13 @@ func (svc *builtIn) Reconfigure(
 		deleteEveryNthValue = svcConfig.DeleteEveryNthWhenDiskFull
 	}
 
-	// Initialize or add collectors based on changes to the component configurations.
-	newCollectors := make(map[resourceMethodMetadata]*collectorAndConfig)
-	if !svc.captureDisabled {
-		for res, resConfs := range captureConfigs {
-			for _, resConf := range resConfs {
-				if resConf.Method == "" {
-					continue
-				}
-				// Create component/method metadata
-				methodMetadata := data.MethodMetadata{
-					API:        resConf.Name.API,
-					MethodName: resConf.Method,
-				}
-
-				componentMethodMetadata := resourceMethodMetadata{
-					ResourceName:   resConf.Name.ShortName(),
-					MethodMetadata: methodMetadata,
-					MethodParams:   fmt.Sprintf("%v", resConf.AdditionalParams),
-				}
-				_, ok := svc.componentMethodFrequencyHz[componentMethodMetadata]
-
-				// Only log capture frequency if the component frequency is new or the frequency has changed
-				// otherwise we'll be logging way too much
-				if !ok || (ok && resConf.CaptureFrequencyHz != svc.componentMethodFrequencyHz[componentMethodMetadata]) {
-					syncVal := "will"
-					if resConf.CaptureFrequencyHz == 0 {
-						syncVal += " not"
-					}
-					svc.logger.Infof(
-						"capture frequency for %s is set to %.2fHz and %s sync", componentMethodMetadata, resConf.CaptureFrequencyHz, syncVal,
-					)
-				}
-
-				// we need this map to keep track of if state has changed in the configs
-				// without it, we will be logging the same message over and over for no reason
-				svc.componentMethodFrequencyHz[componentMethodMetadata] = resConf.CaptureFrequencyHz
-
-				maxCaptureFileSize := svcConfig.MaximumCaptureFileSizeBytes
-				if maxCaptureFileSize == 0 {
-					maxCaptureFileSize = defaultMaxCaptureSize
-				}
-				if !resConf.Disabled && (resConf.CaptureFrequencyHz > 0 || svc.maxCaptureFileSize != maxCaptureFileSize) {
-					// We only use service-level tags.
-					resConf.Tags = svcConfig.Tags
-
-					maxFileSizeChanged := svc.maxCaptureFileSize != maxCaptureFileSize
-					svc.maxCaptureFileSize = maxCaptureFileSize
-
-					newCollectorAndConfig, err := svc.initializeOrUpdateCollector(res, componentMethodMetadata, resConf, maxFileSizeChanged)
-					if err != nil {
-						svc.logger.CErrorw(ctx, "failed to initialize or update collector", "error", err)
-					} else {
-						newCollectors[componentMethodMetadata] = newCollectorAndConfig
-					}
-				}
-			}
-		}
+	if !svcConfig.CaptureDisabled {
 	} else {
 		svc.fileDeletionRoutineCancelFn = nil
 		svc.fileDeletionBackgroundWorkers = nil
 	}
 
-	// If a component/method has been removed from the config, close the collector.
-	svc.collectorsMu.Lock()
-	for md, collAndConfig := range svc.collectors {
-		if _, present := newCollectors[md]; !present {
-			collAndConfig.Collector.Close()
-		}
-	}
-	svc.collectors = newCollectors
-	svc.collectorsMu.Unlock()
-	svc.additionalSyncPaths = svcConfig.AdditionalSyncPaths
+	svc.additionalSyncPaths = append([]string{svc.captureManager.CaptureDir()}, svcConfig.AdditionalSyncPaths...)
 
 	fileLastModifiedMillis := svcConfig.FileLastModifiedMillis
 	if fileLastModifiedMillis <= 0 {
@@ -641,15 +352,16 @@ func (svc *builtIn) Reconfigure(
 			svc.closeSyncer()
 		}
 	}
+
 	// if datacapture is enabled, kick off a go routine to check if disk space is filling due to
 	// cached datacapture files
-	if !svc.captureDisabled {
+	if !svcConfig.CaptureDisabled {
 		fileDeletionCtx, cancelFunc := context.WithCancel(context.Background())
 		svc.fileDeletionRoutineCancelFn = cancelFunc
 		svc.fileDeletionBackgroundWorkers = &sync.WaitGroup{}
 		svc.fileDeletionBackgroundWorkers.Add(1)
 		go pollFilesystem(fileDeletionCtx, svc.fileDeletionBackgroundWorkers,
-			svc.captureDir, deleteEveryNthValue, svc.syncer, svc.logger)
+			svc.captureManager.CaptureDir(), deleteEveryNthValue, svc.syncer, svc.logger)
 	}
 
 	return nil
@@ -730,15 +442,23 @@ func isOffline() bool {
 	return err != nil
 }
 
-func (svc *builtIn) sync(ctx context.Context) {
-	svc.flushCollectors()
-	// Lock while retrieving any values that could be changed during reconfiguration of the data
-	// manager.
+func (svc *builtIn) sync() {
+	svc.captureManager.FlushCollectors()
+
 	svc.lock.Lock()
-	captureDir := svc.captureDir
-	fileLastModifiedMillis := svc.fileLastModifiedMillis
-	additionalSyncPaths := svc.additionalSyncPaths
-	if svc.syncer == nil {
+	if svc.syncer != nil {
+		var toSync []string
+		for _, ap := range svc.additionalSyncPaths {
+			toSync = append(toSync, getAllFilesToSync(ap, svc.fileLastModifiedMillis)...)
+		}
+		syncer := svc.syncer
+		svc.lock.Unlock()
+
+		stopAfter := time.Now().Add(time.Duration(svc.syncIntervalMins * float64(time.Minute)))
+		for _, p := range toSync {
+			syncer.SyncFile(p, stopAfter)
+		}
+	} else {
 		svc.lock.Unlock()
 		return
 	}
@@ -792,39 +512,6 @@ func getAllFilesToSync(ctx context.Context, dirs []string, lastModifiedMillis in
 	}
 }
 
-// Build the component configs associated with the data manager service.
-func (svc *builtIn) updateDataCaptureConfigs(
-	resources resource.Dependencies,
-	conf resource.Config,
-	captureDir string,
-) (map[resource.Resource][]datamanager.DataCaptureConfig, error) {
-	resourceCaptureConfigMap := make(map[resource.Resource][]datamanager.DataCaptureConfig)
-	for name, assocCfg := range conf.AssociatedAttributes {
-		associatedConf, err := utils.AssertType[*datamanager.AssociatedConfig](assocCfg)
-		if err != nil {
-			return nil, err
-		}
-
-		res, err := resources.Lookup(name)
-		if err != nil {
-			svc.logger.Debugw("failed to lookup resource", "error", err)
-			continue
-		}
-
-		captureCopies := make([]datamanager.DataCaptureConfig, len(associatedConf.CaptureMethods))
-		for _, method := range associatedConf.CaptureMethods {
-			method.CaptureDirectory = captureDir
-			captureCopies = append(captureCopies, method)
-		}
-		resourceCaptureConfigMap[res] = captureCopies
-	}
-	return resourceCaptureConfigMap, nil
-}
-
-func generateMetadataKey(component, method string) string {
-	return fmt.Sprintf("%s/%s", component, method)
-}
-
 func pollFilesystem(ctx context.Context, wg *sync.WaitGroup, captureDir string,
 	deleteEveryNth int, syncer datasync.Manager, logger logging.Logger,
 ) {
@@ -863,59 +550,4 @@ func pollFilesystem(ctx context.Context, wg *sync.WaitGroup, captureDir string,
 			}
 		}
 	}
-}
-
-func logCaptureDirSize(ctx context.Context, captureDir string, wg *sync.WaitGroup, logger logging.Logger,
-) {
-	t := clock.Ticker(captureDirSizeLogInterval)
-	defer t.Stop()
-	defer wg.Done()
-	for {
-		if err := ctx.Err(); err != nil {
-			if !errors.Is(err, context.Canceled) {
-				logger.Errorw("data manager context closed unexpectedly", "error", err)
-			}
-			return
-		}
-		select {
-		case <-ctx.Done():
-			return
-		case <-t.C:
-			numFiles := countCaptureDirFiles(ctx, captureDir)
-			if numFiles > minNumFiles {
-				logger.Infof("Capture dir contains %d files", numFiles)
-			}
-		}
-	}
-}
-
-func countCaptureDirFiles(ctx context.Context, captureDir string) int {
-	numFiles := 0
-	//nolint:errcheck
-	_ = filepath.Walk(captureDir, func(path string, info os.FileInfo, err error) error {
-		if ctx.Err() != nil {
-			return filepath.SkipAll
-		}
-		//nolint:nilerr
-		if err != nil {
-			return nil
-		}
-
-		// Do not count the files in the corrupted data directory.
-		if info.IsDir() && info.Name() == datasync.FailedDir {
-			return filepath.SkipDir
-		}
-
-		if info.IsDir() {
-			return nil
-		}
-		// this is intentionally not doing as many checkas as getAllFilesToSync because
-		// this is intended for debugging and does not need to be 100% accurate.
-		isCompletedCaptureFile := filepath.Ext(path) == datacapture.FileExt
-		if isCompletedCaptureFile {
-			numFiles++
-		}
-		return nil
-	})
-	return numFiles
 }
